\name{countgmifs-package}
\alias{countgmifs-package}
\docType{package}
\title{
\packageTitle{countgmifs}: Discrete Response Generalized Monotone Incremental Forward Stagewise Regression
}
\description{
This package provides a function that fits a Poisson or negative binomial model when the number of parameters exceeds the sample size, using the the generalized monotone incremental forward stagewise method.
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{countgmifs}
\packageIndices{countgmifs}
This package contains functions for fitting a penalized discrete response model (either negative binomial or Poisson) and extracting estimated coefficients, predictions, and plots. The model and methods can be used when the response to be predicted is discrete, and is particularly relevant when there are more covariates than observations.
}
\author{
\packageAuthor{countgmifs} Kellie J. Archer <archer.43@osu.edu>

Maintainer: \packageMaintainer{countgmifs} Kellie J. Archer <archer.43@osu.edu>
}
\references{
Makowski M., Archer K.J. (2015) Generalized monotone incremental forward stagewise method for modeling count data: application predicting micronuclei frequency. \emph{Cancer Informatics}, 14(Suppl 2), 97--105.
}
\keyword{ package }
\keyword{ models }
\keyword{ regression }
