% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.countgmifs.R
\name{plot.countgmifs}
\alias{plot.countgmifs}
\title{Plot Solution Path for a Count GMIFS Fitted Model.}
\usage{
\method{plot}{countgmifs}(x, type = "trace", xlab = NULL, ylab = NULL,
  main = NULL, ...)
}
\arguments{
\item{x}{a \code{countgmifs} object.}

\item{type}{default is \code{"trace"} which plots the coefficient path for the fitted object. Also available are \code{"AIC"}, \code{"BIC"}, and \code{"logLik"}.}

\item{xlab}{a default x-axis label will be used which can be changed by specifying a user-defined x-axis label.}

\item{ylab}{a default y-axis label will be used which can be changed by specifying a user-defined y-axis label.}

\item{main}{a default main title will be used which can be changed by specifying a user-defined main title.}

\item{...}{other arguments.}
}
\description{
This function plots either the coefficient path, the AIC, or the log-likelihood for a fitted \code{countgmifs} object.
}
\seealso{
See Also \code{\link{countgmifs}}, \code{\link{coef.countgmifs}}, \code{\link{summary.countgmifs}}, \code{\link{predict.countgmifs}}
}
\keyword{methods}
