% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.countgmifs.R
\name{print.countgmifs}
\alias{print.countgmifs}
\title{Print the Contents of a Count GMIFS Fitted Object.}
\usage{
\method{print}{countgmifs}(x, ...)
}
\arguments{
\item{x}{an \code{countgmifs} fitted object.}

\item{...}{other arguments.}
}
\description{
This function prints the names of the list objects from an \code{countgmifs} fitted model
}
\seealso{
See Also \code{\link{countgmifs}}, \code{\link{coef.countgmifs}}, \code{\link{summary.countgmifs}}, \code{\link{plot.countgmifs}}
}
\keyword{methods}
